requirements.txt
alembic==1.12.1
anyio==3.7.1
attrs==23.1.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
click==8.1.7
dnspython==2.4.2
docstring-parser==0.15
email-validator==2.1.0.post1
Faker==19.13.0
fastapi==0.110.0
FastAPI-SQLAlchemy==0.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==23.2
pandas==2.2.2
passlib==1.7.4
pluggy==1.4.0
psycopg2-binary==2.9.9
py==1.11.0
pycparser==2.21
pydantic==2.6
pydantic-settings==2.1.0
PyJWT==2.8.0
pyparsing==3.1.2
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.36.3
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
zipp==3.18.1
cloudinary==1.44.0